seperate texture code into module
